rename tools and make the report_tool_infos.py exec, can copy it to /usr/local/bin/report_tool_infos
